# docs/swagger.yaml
basePath: /api
definitions:
  models.Post:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@backend.local
    name: API Support
  description: |-
    REST API for posts, users, and image uploads.

    Auth: JWT access token (24h) + rotating refresh token (30d).
    Roles: user, admin, superadmin. Admin+ can moderate users.
    All responses include a request_id for log correlation.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Backend API
  version: "1.0"
paths:
...
  /posts:
    get:
      parameters:
      - description: Page size (max 100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Filter by owner UUID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      parameters:
      - description: Post
        in: body
        name: body
        required: true
        schema:
          properties:
            body:
              type: string
            image_url:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a post
      tags:
      - posts
  /posts/{id}:
    delete:
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get one post
      tags:
      - posts
    put:
      consumes:
      - application/json
      parameters:
      - description: Post UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields
        in: body
        name: body
        required: true
        schema:
          properties:
            body:
              type: string
            image_url:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...